Add Section 30C guidance note to infoPanel.
At request from Katrina McLaughlin on 2024-01-24, the "homepage" of the
map has updated language about the Section 30C guidance announcement.

Added the following paragraph:

"""
On January 19th, 2024, the Treasury Department and the Internal Revenue Service issued guidance on eligible census tracts for the alternative fuel refueling property credit (§30C). We will be updating the mapper to reflect this new guidance.

"""